requirements.txt
# Project requirements
# Usage: pip install -r requirements.txt

requests==2.30.0
pytest==7.3.1
allure-pytest==2.13.2
selenium==4.9.1
webdriver-manager==3.8.6
Faker==18.9.0
